move relevant methods into domain specific files, add tests
