+++ src/control/client/connect.go
@@ -194,84 +194,3 @@ func (c *connList) makeRequests(
 
 	return cMap
 }
-
-// storageResult container for results from multiple storage subsystems queries.
-type StorageResult struct {
-	nvme nvmeResult
-	scm  scmResult
-}
-
-// listStorageRequest is to be called as a goroutine and returns result
-// containing locally attached NVMe and SCM storage devices over channel.
-func listStorageRequest(mc Control, ch chan result) {
-	sRes := StorageResult{}
-
-	ctrlrs, err := mc.listNvmeCtrlrs()
-	sRes.nvme = nvmeResult{ctrlrs, err}
-
-	mms, err := mc.listScmModules()
-	sRes.scm = scmResult{mms, err}
-
-	ch <- result{mc.getAddress(), sRes, nil} // result.err is ignored
-}
-
-// ListStorage returns locally-attached nonvolatile storage devices for each
-// connected server.
-func (c *connList) ListStorage() (cNvmeMap, cScmMap) {
-	cResults := c.makeRequests(listStorageRequest)
-	cCtrlrs := make(cNvmeMap) // mapping of server address to NVMe SSDs
-	cModules := make(cScmMap) // mapping of server address to SCM modules
-
-	for _, res := range cResults {
-		// we want to extract obj regardless of error as may only refer
-		// to one of the subsystems, ignore res.err
-		storageRes, ok := res.value.(StorageResult)
-		if !ok {
-			err := fmt.Errorf(
-				"type assertion failed, wanted %+v got %+v",
-				StorageResult{}, res.value)
-
-			cCtrlrs[res.address] = nvmeResult{nil, err}
-			cModules[res.address] = scmResult{nil, err}
-			continue
-		}
-
-		cCtrlrs[res.address] = storageRes.nvme
-		cModules[res.address] = storageRes.scm
-	}
-
-	return cCtrlrs, cModules
-}
-
-// KillRank Will terminate server running at given rank on pool specified by
-// uuid.
-func (c *connList) KillRank(uuid string, rank uint32) ResultMap {
-	errors := make(ResultMap)
-	ch := make(chan result)
-
-	for _, mc := range c.controllers {
-		go func() {
-			ch <- result{
-				mc.getAddress(),
-				nil,
-				mc.killRank(uuid, rank),
-			}
-		}()
-	}
-
-	for range c.controllers {
-		res := <-ch
-		errors[res.address] = res
-	}
-
-	return errors
-}
-
-// NewConnect is a factory for Connect interface to operate over
-// multiple clients.
-func NewConnect() Connect {
-	return &connList{
-		factory:     &controllerFactory{},
-		controllers: []Control{},
-	}
-}
